requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
numpy
opencv-python
pyaudio
openai-whisper
sentencepiece
transformers==4.42.3
huggingface_hub
accelerate
sacremoses
Pillow

torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128